+++ examples/resources/statusdrift_sla_policy/resource.tf
@@ -2,6 +2,7 @@ resource "statusdrift_sla_policy" "production" {
   name          = "Production SLA"
   uptime_target = 99.9
   window_type   = "calendar_month"
+  scope         = "org"
 
   response_time_sla_enabled  = true
   response_time_threshold_ms = 500